refactor: service 추가에 따른 테스트 로직 구조 변환
